feat: implement PrimaryAction component and enhance SmartBtn with new styles and functionality
